requirements.txt
gunicorn==19.9.0
itsdangerous==1.1.0
MarkupSafe==1.1.1
Flask==1.1.2
PIL==8.0.0
imageio==2.9.0
scipy==1.5.2
numpy==1.17.2
tensorflow-gpu==2.3.0
werkzeug==0.12.2
scipy==0.19.0
Jinja2==2.11.2
opencv-python-headless==4.2.0.32
